+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Migrate firmware inventory information decoder to stream API
 - Refactor logging engine
 
 ### Fixed